Update fat image, add hw version text.
* Fat image now had updated version of GodMode9i. (Version with slot2
support that now has latest commits from main branch of GM9i)
* Autoboot can now be toggled on/off via single const bool. Compile with
the bool set to false if you wish to disable autoboot.
* New card command added to read hw version from EZP cart. This command
was found courtasy of Lifehackerhansol. UI code updated to dispaly this
in top right corner of top screen UI. (This is only displayed if file
browser UI is activated)